refactor(price-converter): group rate providers
