+++ apps/price-converter/src/rates/bitpay/fetchBitpayRates.test.ts
