+++ apps/price-converter/src/rates/bitpay/fixtures/bitpay.json
