+++ apps/price-converter/src/rates/blockchainInfo/BlockchainInfoResponse.ts
@@ -1,0 +1,14 @@
+import { String as EvoluString, object, record, Unknown } from '@evolu/common';
+import { PositiveFiniteNumber } from '../positiveFiniteNumber.js';
+
+const UnknownProperties = record(EvoluString, Unknown);
+
+const BlockchainInfoRate = object(
+    {
+        last: PositiveFiniteNumber,
+    },
+    UnknownProperties,
+);
+
+export const BlockchainInfoResponse = record(EvoluString, BlockchainInfoRate);
+export type BlockchainInfoResponse = typeof BlockchainInfoResponse.Output;
