+++ apps/price-converter/src/rates/coingecko/fetchCoingeckoRates.test.ts
